Organize code into new files
Seperate out the different parts of the code into seperate files. All
game states will go in the src/states directory, game objects will be
sorted into different folders based on the use. Other scripts, a
movement controller, for example, will be in the src/components
directory.

Assets will be in the assets directory, under a descriptive
subdirectory. For example, the player texture would be undere
assets/player.
