+++ src/index.ts
@@ -1,5 +1,5 @@
 import { MerlinEngine } from "merlin-game-engine/dist";
-import { TestGame } from "./test";
+import { TestGame } from "./states/testGame";
 
 const engine = new MerlinEngine();
 const game = new TestGame();
