+++ src/types/assets.d.ts
